implement `Cairo::FontFace.quartz_supported?` and use this in test
"Helvetica" is specified in cairo/cairoint.h as follows:

    #define CAIRO_QUARTZ_FONT_FAMILY_DEFAULT  "Helvetica"
this definition is redfined:

	#define CAIRO_FONT_FAMILY_DEFAULT CAIRO_QUARTZ_FONT_FAMILY_DEFAULT

And used in `_cairo_gstate_ensure_font_face()`

    font_face = cairo_toy_font_face_create (CAIRO_FONT_FAMILY_DEFAULT,
                                            CAIRO_FONT_SLANT_DEFAULT,
                                            CAIRO_FONT_WEIGHT_DEFAULT);

and `_cairo_gstate_ensure_font_face()` is used in
`cairo_get_font_face()`.

So it is suitable `Cairo::FontFace.quartz_supported?` definition place
in Cairo::FontFace, isn't it?

implement `quartz?` in CairoTestUtils